+++ App/src/analytics/TimerEvent.ts
@@ -1,15 +1,27 @@
-import { trackEvent } from './CognitoAnalytics';
+import { IAnalyticsProvider } from './IAuditor';
+import { logger } from "./logger";
 import { Metrics, Params } from './Types';
 
 export class TimerEvent {
     start: number;
-    constructor(private event: string) {
+
+    constructor(
+        private provider: IAnalyticsProvider,
+        private event: string) {
         this.start = Date.now();
     }
+
     public submit(params?: Params, metrics?: Metrics) {
-        trackEvent(this.event, params, {
-            ...(metrics || {}),
-            duration: Date.now() - this.start,
-        });
+        if (!this.provider) { return; }
+
+        try {
+            this.provider.trackEvent(this.event, params, {
+                ...(metrics || {}),
+                duration: Date.now() - this.start,
+            });
+        }
+        catch (e) {
+            logger.error(e, "trackAction failed");
+        }
     }
 }